+++ src/Providers/ConsoleServiceProvider.php
@@ -58,6 +58,7 @@ class ConsoleServiceProvider extends ServiceProvider
             Commands\Make\ControllerMakeCommand::class,
             Commands\Make\EventMakeCommand::class,
             Commands\Make\EnumMakeCommand::class,
+            Commands\Make\ExceptionMakeCommand::class,
             Commands\Make\FactoryMakeCommand::class,
             Commands\Make\HelperMakeCommand::class,
             Commands\Make\JobMakeCommand::class,