+++ src/Providers/ConsoleServiceProvider.php
@@ -77,6 +77,7 @@ class ConsoleServiceProvider extends ServiceProvider
             Commands\Make\ResourceMakeCommand::class,
             Commands\Make\RouteProviderMakeCommand::class,
             Commands\Make\RuleMakeCommand::class,
+            Commands\Make\ScopeMakeCommand::class,
             Commands\Make\SeedMakeCommand::class,
             Commands\Make\ServiceMakeCommand::class,
             Commands\Make\TraitMakeCommand::class,