+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All Notable changes to `laravel-modules` will be documented in this file.
 ## Next
 
 - [@dcblogdev](https://github.com/dcblogdev) added make-action command
+- [@dcblogdev](https://github.com/dcblogdev) added make-cast command
 - [@dcblogdev](https://github.com/dcblogdev) added make-enum command
 - [@dcblogdev](https://github.com/dcblogdev) added make-helper command
 - [@dcblogdev](https://github.com/dcblogdev) added make-trait command